+++ base/iterators.jl
@@ -10,7 +10,8 @@ import Base: start, done, next, isempty, length, size, eltype, iteratorsize, ite
 using Base: tail, tuple_type_head, tuple_type_tail, tuple_type_cons, SizeUnknown, HasLength, HasShape,
             IsInfinite, EltypeUnknown, HasEltype, OneTo, @propagate_inbounds, Generator, AbstractRange
 
-export enumerate, zip, rest, countfrom, take, drop, cycle, repeated, product, flatten, partition
+export enumerate, zip, rest, countfrom, take, drop, cycle, repeated, product, flatten,
+       partition, only
 
 _min_length(a, b, ::IsInfinite, ::IsInfinite) = min(length(a),length(b)) # inherit behaviour, error
 _min_length(a, b, A, ::IsInfinite) = length(a)
@@ -913,5 +914,33 @@ function next(itr::PartitionIterator, state)
     end
     return resize!(v, i), state
 end
+
+"""
+    only(x)
+
+Returns the one and only element of collection `x`, and throws an error if the collection
+has zero or multiple elements.
+"""
+Base.@propagate_inbounds function only(x)
+    i = start(x)
+    @boundscheck if done(x, i)
+        error("Collection is empty, must contain exactly 1 element")
+    end
+    (ret, i) = next(x, i)
+    @boundscheck if !done(x, i)
+        error("Collection has multiple elements, must contain exactly 1 element")
+    end
+    return ret
+end
+
+# Collections of known size
+only(x::Tuple{}) = error("Tuple is empty, must contain exactly 1 element")
+only(x::Tuple{Any}) = x[1]
+only(x::Tuple) = error("Tuple contains $(length(x)) elements, must contain exactly 1 element")
+
+only(a::AbstractArray{<:Any, 0}) = @inbounds return a[]
+
+only(x::NamedTuple{<:Any, <:Tuple{Any}}) = first(x)
+only(x::NamedTuple) = error("NamedTuple contains $(length(x)) elements, must contain exactly 1 element")
 
 end